auto_gpt_agent::models::agents::agent_traits
Trait SpecialFunctions

pub trait SpecialFunctions: Debug {
    // Required methods
    fn get_attributes_from_agent(&self) -> &BasicAgent;
    fn execute<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        fact_sheet: &'life1 mut FactSheet,
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Defines the interface for agents to implement special functions.

Methods

get_attributes_from_agent: Returns a reference to the agent’s attributes.

execute: Executes the agent’s workflow using a mutable reference to the FactSheet.

Required Methods

fn get_attributes_from_agent(&self) -> &BasicAgent

fn execute<'life0, 'life1, 'async_trait>( &'life0 mut self, fact_sheet: &'life1 mut FactSheet, ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,